Fix for target name floating after UnitPanel hiding
Fix for #145. The UnitPanels don't use :Hide() to hide the frames, instead the alpha is changed on the textures. When no target is selected, the text gets set to an empty string. The problem here is: when teleporting (between the events "PLAYER_LEAVING_WORLD" and "PLAYER_ENTERING_WORLD") the target changed event won't trigger. This commit adds "PLAYER_ENTERING_WORLD" to the unitname's eventhandler.